MYyyyyYyyYyThhhriillllll
To make advanced tribal armor you now need mythril which can be found in
the ash walkers nest inside of a Necropolis chest. There is a plentiful
amount so a miner willing to assail a nest will find plentiful loots.
Ashwalkers also spawn with 2 toxin kits to compensate for weaver venom,
this means they have a chance at saving a poisoned tribesman. Made it so
all meds are in one crate to free up space, moved the IV drip next to
the OR table and expanded the right side a bit to act as a crafting
area.