Adds in attack verbs to 4 items
The station charter, cyborg endoskeleton, all ID cards and the grey
jumpsuit get unique attack verbs. The endoskeleton also plays the sound
when you borg someone ("I live.. again") when you hit someone or
yourself with it(this doesn't mean you go on a spree of hitting people
with the endoskeleton; it injures, though it would be a cheesy death if
you get beaten to death by an endoskeleton) as well. The attack verbs
for the ID cards varies on its type(example; hitting people with an
abductor ID may say alienated, abducted, etc.). The grey jumpsuit attack
verb is self-explanatory.