Disarming a live nuclear Bomb
This was blocked previously by the snowflake coded "nuke screwdriver" that was identical in every way to the regular screwdriver except for this tiny change. Now the crew has an incentive to try to stop the live nuke before it kills everyone. They'll struggle with the massive bursts of radiation, though.