Fixes Suit Storage Units losing auxiliary slots
This fixes Suit Storage Units losing their auxiliary slots when a human is
placed inside of and then removed from a Suit Storage Unit or the items
inside are incinerated. Instead of setting extra_items to null when
drop_contents is called, it now just sets it to an empty list.