Fixes mechs w/o cult occupant opening cult doors
This fixes mechs opening doors they are not allowed to open if they happen
to posses the required access for the door or the door is in emergency
mode. This bug allowed mechs with a non-cultist occupant to open a cult
door because cult doors do not have a list of required accesses.